Qt6 is stricter when casting numericals and chars. Add an explicit cast/conversion to better digest them
cast

+++ src/modules/list/ListWindow.cpp
@@ -124,7 +124,7 @@ QSize ChannelTreeWidgetItemDelegate::sizeHint(const QStyleOptionViewItem & sovIt
 			break;
 		case 1:
 			//users
-			return QSize(fm.horizontalAdvance(item->itemData()->m_szUsers.toInt()), iHeight);
+			return QSize(fm.horizontalAdvance(item->itemData()->m_szUsers), iHeight);
 			break;
 		case 2:
 		default:
